+++ src/panels/ribbon_tools.cpp
@@ -285,8 +285,8 @@ void RibbonPanel::OnDropDown(wxRibbonToolBarEvent& event)
         case NewScrolled:
             {
                 wxMenu menu;
-                menu.Append(gen_wxScrolledWindow, "Insert wxScrolled<wxPanel> (wxScrolledWindow)");
-                menu.Append(gen_wxScrolledCanvas, "Insert wxScrolled<wxWindow> (wxScrolledCanvas)");
+                menu.Append(gen_wxScrolledWindow, "Insert wxScrolledWindow");
+                menu.Append(gen_wxScrolledCanvas, "Insert wxScrolledCanvas");
 
                 menu.Bind(wxEVT_MENU, &RibbonPanel::OnMenuEvent, this, wxID_ANY);
                 event.PopupMenu(&menu);