+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 - Fixed code generation for Python and Ruby wxStdDialogButtonSizer events
 - Ruby code generation no longer writes the class `end` line before the final generated comment block
 - You can add a menubar or toolbar to a wxFrame that was created without them
+- Fixed inconsistent generation of wxScrolledWindow versus wxScrolled<wxPanel> -- wxScrolledWindow is now always used.
 
 ## [Released (1.2.1)]